[Add] Initial implementation of *nix MAC getter
Use the ioctl system calls from *nix to get the HW address.
The code is taken from the man page and adjusted for modern C++.

The implementation is a bit long and may be simplified.
All UTs pass with this change on Fedora 29, gcc 8.3.0.

This might even work on __APPLE__, but have not tested it.